Move SqlDataSet class from plan_conversion to dataset package
The other DataSet classes are all in one place, and this is no
longer strictly a plan_conversion artifact (indeed, it never
was, it just looked that way due to the generics).
