+++ metricflow/dataset/semantic_model_adapter.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from dbt_semantic_interfaces.references import SemanticModelReference
 
+from metricflow.dataset.sql_dataset import SqlDataSet
 from metricflow.instances import InstanceSet
-from metricflow.plan_conversion.sql_dataset import SqlDataSet
 from metricflow.sql.sql_plan import SqlSelectStatementNode
